pyproject.toml
[project]
name = "hedorah-agent"
version = "0.1.0"
description = "Personal research agent for AI interpretability - PDF processing, paper analysis, and experiment design"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "anthropic>=0.40.0",
    "watchdog>=6.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
hedorah = "hedorah.cli:main"

[tool.uv]
package = true
